requirements.txt
aniso8601==8.0.0
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
Flask-RESTful==0.3.8
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
post==2019.4.13
protobuf==3.6.1
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
six==1.14.0
Werkzeug==1.0.0
